Add eslint config, eslint the things
We don't have an official JS style guide, so I just went with Google's to
augment the eslint recommended lints.  Feel free to push back on that.

The changes to `index.js` and `test/` were done by running:

    ./node_modules/.bin/eslint --fix index.js test/

A few errors (use of `snake_case` variables) had to be fixed by hand, though.